dashboard/requirements.txt
streamlit>=1.32
plotly>=5.18
supabase>=2.3
pandas>=2.0
python-dotenv>=1.0
